Fix flay TopicHandlerTest#testCreateTopicByProbability
